+++ js/comment.js
@@ -1,0 +1,20 @@
+import { getRandomInt } from './randomUtils.js';
+
+const messages = [
+    "Всё отлично!",
+    "В целом всё неплохо. Но не всё.",
+    "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+    "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+    "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+    "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
+];
+const names = ["Артём", "Мария", "Иван", "Светлана", "Алексей", "Дарья"];
+
+export function createComment(id) {
+    return {
+        id: id,
+        avatar: `img/avatar-${getRandomInt(1, 6)}.svg`,
+        message: messages[getRandomInt(0, messages.length - 1)],
+        name: names[getRandomInt(0, names.length - 1)]
+    };
+}
